# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Architectural drawing management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      tags: [storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for a drawing file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Invalid metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /storage/drawings/{id}/drive-upload:
    post:
      operationId: uploadDrawingToDrive
      tags: [storage]
      summary: Upload a drawing file to Google Drive
      parameters:
        - $ref: "#/components/parameters/Id"
        - name: X-File-Name
          in: header
          required: true
          schema:
            type: string
        - name: X-File-Content-Type
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Recorded Drive upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DrawingUpload"
        "409":
          description: Google Drive is not connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /storage/objects/{objectPath}:
    get:
      tags: [storage]
      operationId: getStorageObject
      summary: Serve a drawing file
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Drawing file content
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /storage/drive-files/{fileId}:
    get:
      operationId: getDriveFile
      tags: [storage]
      summary: Stream a drawing file from Google Drive
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Drawing file content
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: number
        minimum: 1
...
  schemas:
...
    DrawingUpload:
      type: object
      properties:
        id:
          type: number
        drawingId:
          type: number
        filePath:
          type: string
        fileName:
          type: string
        fileSize:
          type: number
        contentType:
          type: string
        uploadedBy:
          type: string
        uploadedAt:
          type: string
          format: date-time
      required: [id, drawingId, filePath, fileName, fileSize, contentType, uploadedBy, uploadedAt]
...
    UploadUrlRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: number
          minimum: 1
        contentType:
          type: string
          minLength: 1
      required: [name, size, contentType]
    UploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
      required: [uploadURL, objectPath, metadata]
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: string
      required: [error]